Move the PostSettingsModal to the modals directory
